# docs/swagger.yaml
basePath: /1sat
definitions:
...
  admin.ZSetItem:
    properties:
      score:
        type: number
      value:
        type: string
    type: object
  admin.ZSetResponse:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/admin.ZSetItem'
        type: array
      key:
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/b-open-io/1sat-stack
  description: Composable BSV blockchain services API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 1Sat Stack API
  version: "1.0"
paths:
  /admin/blacklist:
    get:
      description: Returns the list of blacklisted topics
      produces:
      - application/json
      responses:
        "200":
          description: List of blacklisted topics
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get blacklist
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Adds a topic to the blacklist
      parameters:
      - description: Topic to add
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add to blacklist
      tags:
      - admin
  /admin/blacklist/{topic}:
    delete:
      description: Removes a topic from the blacklist
      parameters:
      - description: Topic ID to remove
        in: path
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove from blacklist
      tags:
      - admin
...
  /admin/progress/{id}:
    delete:
      description: Deletes a sync progress entry
      parameters:
      - description: Progress ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete progress
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Updates a sync progress entry
      parameters:
      - description: Progress ID
        in: path
        name: id
        required: true
        type: string
      - description: Block height
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update progress
      tags:
      - admin
...
  /admin/whitelist:
    get:
      description: Returns the list of whitelisted BSV21 tokens (always active)
      produces:
      - application/json
      responses:
        "200":
          description: List of whitelisted tokens
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get whitelist
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Adds a BSV21 token to the whitelist (always active)
      parameters:
      - description: Token to add
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add to whitelist
      tags:
      - admin
  /admin/whitelist/{token}:
    delete:
      description: Removes a BSV21 token from the whitelist
      parameters:
      - description: Token ID to remove
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove from whitelist
      tags:
      - admin
  /admin/zset/{key}:
    get:
      description: Returns the first 25 items from any sorted set by key
      parameters:
      - description: Sorted set key (e.g., q:tok:abc123, z:ev:own:address)
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sorted set info and items
          schema:
            $ref: '#/definitions/admin.ZSetResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get sorted set items
      tags:
      - admin
...
swagger: "2.0"